backend_python/RegistroDeHorasApp-backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
gunicorn==20.1.0
whitenoise==6.5.0
psycopg2-binary==2.9.6
python-decouple==3.8
django-filter==23.2
dj-database-url==2.0.0 